Fixed code to remove game warning

+++ sources/game.cpp
@@ -32,7 +32,7 @@ void Game::play()
 //kept outside of 'Game' to free up stack from holding 'words' while playing
 string Game::getRandWord()
 {
-    srand ( time(NULL) ); // initialize the random seed
+    srand ( static_cast<unsigned int>(time(NULL)) ); // initialize the random seed
     vector<string> words = {"obama", "trump", "republican", "democrat", "donkey",
         "monkey", "computer", "commuter", "screensaver", "random", "anagram"
     };
